refactor db viewer
