+++ src/features/DevPanel/PostgresViewer/DataTable/index.tsx
@@ -3,7 +3,11 @@ import { Button } from 'antd';
 import { createStyles } from 'antd-style';
 import { Download, Filter, RefreshCw } from 'lucide-react';
 import React from 'react';
+import { mutate } from 'swr';
 
+import { exportService } from '@/services/export';
+
+import { FETCH_TABLE_DATA_KEY } from '../usePgTable';
 import Table from './Table';
 
 const useStyles = createStyles(({ token, css }) => ({
@@ -53,8 +57,21 @@ const DataTable = ({ tableName }: DataTableProps) => {
           <Button color={'default'} icon={<Icon icon={Filter} />} variant={'filled'}>
             Filter
           </Button>
-          <ActionIcon icon={Download} title={'Export'} />
-          <ActionIcon icon={RefreshCw} title={'Refresh'} />
+          <ActionIcon
+            icon={Download}
+            onClick={async () => {
+              const data = await exportService.exportData();
+              console.log(data);
+            }}
+            title={'Export'}
+          />
+          <ActionIcon
+            icon={RefreshCw}
+            onClick={async () => {
+              await mutate(FETCH_TABLE_DATA_KEY(tableName));
+            }}
+            title={'Refresh'}
+          />
         </div>
       </div>
